Extract new logging functions to Carto::Logger
